Added category and items to the correct folder and import the static module from django.contrib to be able to read MEDIA_URL
