+++ .gitignore
@@ -8,3 +8,4 @@ db.sqlite3
 __pycache__
 env
 migrations
+media
