+++ core/admin.py
@@ -1,5 +1,1 @@
-from django.contrib import admin
-from.models import Category, Item
 
-admin.site.register(Category)
-admin.site.register(Item)
