+++ core/templates/core/index.html
@@ -3,5 +3,40 @@
 {% block title %}Welcome{% endblock %}
 
 {% block content%}
-    <h1>The front page</h1>
+<div class="mt-6 px-6 py-12 bg-gray-100 rounded-xl">
+    <h2 class="mb-12 text-2xl text-center">Newest items</h2>
+
+    <div class="grid grid-cols-3 gap-3">
+
+        {% for item in items %}
+        <div>
+            <a href="#">
+                <div>
+                    <img src="{{ item.image.url }}" class="rounded-t-xl">
+                </div>
+
+                <div class="p-6 bg-white rounded-b-xl">
+                    <h2 class="text-2xl">{{ item.name }}</h2>
+                    <p class="text-gray-500">Price: {{ item.price }}</p>
+                </div>
+            </a>
+        </div>
+        {% endfor %}
+    </div>
+</div>
+
+<div class="mt-6 px-6 py-12 bg-gray-100 rounded-xl">
+    <h2 class="mb-12 text-2xl text-center">Categories</h2>
+
+    <div class="grid grid-cols-3 gap-3">
+        <div>
+            <a href="#">
+                <div class="p-6 bg-white rounded-b-xl">
+                    <h2 class="text-2xl">name</h2>
+                    <p class="text-gray-500"> items</p>
+                </div>
+            </a>
+        </div>
+    </div>
+</div>
 {% endblock %}
