+++ core/views.py
@@ -1,8 +1,14 @@
 from django.shortcuts import render
+from item.models import Category, Item
 
 
 def index(request):
-    return render(request, 'core/index.html')
+    items = Item.objects.filter(is_sold=False)[0:6]
+    categories = Category.objects.all()
+    return render(request, 'core/index.html', {
+        'categories': categories,
+        'items': items,
+    })
 
 
 def contact(request):
