+++ item/admin.py
@@ -1,3 +1,7 @@
 from django.contrib import admin
 
-# Register your models here.
+from django.contrib import admin
+from .models import Category, Item
+
+admin.site.register(Category)
+admin.site.register(Item)
